refactor(orchestrator): extract queue + reconcile + lifecycle from runner.go
runner.go was 2234 LOC and 46 funcs across queue, reconcile,
in-flight job tracking, approval lifecycle, event emission, and
artifact upsert. Pull the queue / reconcile / job-tracker /
heartbeat / shutdown cluster into a sibling — the most
concurrency-dense slice of the file lives in one place now.

  runner.go:       2234 → 1829 LOC
  runner_queue.go: new, 434 LOC

What moves to runner_queue.go

  Queue dispatch
  - processQueue
  - processQueuedRun
  - enqueueRun
  - executorForTask

  In-flight job tracker
  - registerJob
  - unregisterJob
  - hasInFlightJob

  Heartbeat + claim lifecycle
  - heartbeatClaim

  Queue API + reconcile loop
  - getQueue, SetQueue
  - ReconcilePendingRuns
  - StartReconcileLoop
  - reconcileStaleRuns

  Process shutdown
  - Shutdown

What stays in runner.go

  Lifecycle (StartTask, ResumeTask, RetryTask, ContinueTask,
  CancelTask and variants), approval-resolved + reject hooks
  (ResumeTaskAfterApproval, RejectTaskAfterApproval,
  cancelPendingApprovalsForRun, recordApprovalResolved, the
  approvalSpecForTask / createApprovalForTask helpers), event
  emission (emitRunEvent), step/artifact upsert (upsertStep,
  upsertArtifact, gitSummaryArtifact, resumeCheckpointForRun,
  finalizeFailedRun), and Runner construction.

Splits of those clusters are the 4b (approvals) + 4c
(events + IO) follow-ups.

The receiver type stays *Runner everywhere; no public API change;
no test changes. Imports per file pruned via goimports.

Verification

  go build ./...                                                — clean
  go vet ./...                                                  — clean
  gofmt -s -l .                                                 — clean
  go test ./internal/orchestrator -run 'TestRunner|runner_*'
    -race -count=5                                              — 50 / 50 pass
  go test ./... -race -count=1                                  — 1779 / 1779 pass
